+++ src/components/activities/RosterPanel.tsx
@@ -1,10 +1,13 @@
+import { Button, Chip, DialogActions, Divider, Table, TableBody, TableCell, TableFooter, TableRow } from '@mui/material';
 import Card from '@mui/material/Card';
 import CardActionArea from '@mui/material/CardActionArea';
 import { PaperProps } from '@mui/material/Paper';
-import { FunctionComponent, ReactNode } from 'react';
+import { useTheme } from '@mui/material/styles';
+import { format as formatDate } from 'date-fns';
+import { FunctionComponent, ReactNode, useEffect, useState } from 'react';
 
-import { Box, Paper, Stack, Typography } from '@respond/components/Material';
-import { Activity, getStatusCssColor, Participant, ParticipantStatus, ParticipatingOrg } from '@respond/types/activity';
+import { Box, Dialog, DialogContent, DialogTitle, Paper, Stack, Typography, useMediaQuery } from '@respond/components/Material';
+import { Activity, getStatusCssColor, getStatusText, isActive, Participant, ParticipantStatus, ParticipantUpdate, ParticipatingOrg } from '@respond/types/activity';
 
 interface RosterPanelProps {
   activity: Activity;
@@ -50,4 +53,97 @@ export function RosterRowCard({ status, children, onClick, ...props }: PaperProp
       {cardContent}
     </Card>
   );
+}
+
+export function ParticipantDialog({ open, participant, activity, onClose }: { open: boolean; onClose: () => void; participant?: Participant; activity: Activity }) {
+  const isMobile = useMediaQuery(useTheme().breakpoints.down('md'));
+
+  if (!participant) return <></>;
+
+  const name = `${participant.firstname} ${participant.lastname}`;
+  return (
+    <Dialog open={open} onClose={onClose}>
+      <DialogTitle style={{ borderBottom: 'solid 4px ' + getStatusCssColor(participant.timeline[0].status) }} alignItems="center" justifyContent="space-between" display="flex">
+        <Box>{name}</Box>
+        <Typography style={{ color: getStatusCssColor(participant.timeline[0].status) }}>{getStatusText(participant.timeline[0].status)}</Typography>
+      </DialogTitle>
+      <DialogContent>
+        <Stack direction={{ xs: 'column', md: 'row' }} spacing={2} sx={{ pt: 2 }} divider={<Divider orientation={isMobile ? 'horizontal' : 'vertical'} flexItem />}>
+          <Box>
+            <img //
+              src={`/api/v1/organizations/${participant.organizationId}/members/${participant.id}/photo`}
+              alt={`Photo of ${name}`}
+              style={{ width: '8rem', minHeight: '10rem', border: 'solid 1px #777', borderRadius: '4px' }}
+            />
+            <Typography fontWeight={600}>{activity.organizations[participant.organizationId].rosterName ?? activity.organizations[participant.organizationId].title}</Typography>
+            <Box>{participant.tags?.map((t) => <Chip sx={{ mr: '3px' }} key={t} label={t} variant="outlined" size="small" />)}</Box>
+          </Box>
+          <Box>
+            <Typography>Timeline:</Typography>
+            <Table size="small">
+              <TableBody>
+                {[...participant.timeline].reverse().map((t) => (
+                  <TableRow key={t.time}>
+                    <TableCell>{activity.organizations[t.organizationId].rosterName ?? activity.organizations[t.organizationId].title}</TableCell>
+                    <TableCell>{getStatusText(t.status)}</TableCell>
+                    <TableCell>{formatDate(t.time, 'EEE yyyy-MM-dd HHmm')}</TableCell>
+                  </TableRow>
+                ))}
+              </TableBody>
+              <TableFooter>
+                <TableRow>
+                  <TableCell colSpan={3} align="right" style={{ borderBottomWidth: 0 }}>
+                    <Typography>
+                      Total Hours: <ParticipantHoursText participant={participant} />
+                    </Typography>
+                    {participant.miles !== undefined && <Typography>Total Miles: {participant.miles}</Typography>}
+                  </TableCell>
+                </TableRow>
+              </TableFooter>
+            </Table>
+          </Box>
+        </Stack>
+        {/* <DialogContentText>Mark this activity as deleted? Any data it contains will stop contributing to report totals.</DialogContentText> */}
+      </DialogContent>
+      <DialogActions>
+        <Button autoFocus onClick={onClose}>
+          Close
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
+
+function isOnClock(status: ParticipantStatus) {
+  if (status === ParticipantStatus.Standby) return false;
+  return isActive(status);
+}
+
+function getLastTimeout(lastUpdate: ParticipantUpdate) {
+  if (!isOnClock(lastUpdate.status)) return lastUpdate.time;
+  return new Date().getTime();
+}
+
+function ParticipantHoursText({ participant }: { participant: Participant }) {
+  const lastTimeline = participant.timeline[0];
+  const [latestTimeout, setLatestTimeout] = useState<number>(getLastTimeout(lastTimeline));
+
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setLatestTimeout(getLastTimeout(lastTimeline));
+    }, 10000);
+    return () => clearTimeout(timer);
+  }, [lastTimeline, latestTimeout, participant.firstname]);
+
+  let timeOnClock = 0;
+  let timeout = latestTimeout;
+  for (const t of participant.timeline) {
+    if (isOnClock(t.status)) {
+      timeOnClock += timeout - t.time;
+    } else {
+      timeout = t.time;
+    }
+  }
+
+  return <>{Math.round(timeOnClock / 1000 / 60 / 15) / 4}</>;
 }
